fix: accept/deny contrib + code lint

+++ src/main/java/com/osa/openstreetart/controller/ContribController.java
@@ -53,7 +53,7 @@ public class ContribController {
    
     @PostMapping(value = "/contrib/{art_id}")
 	public ResponseEntity<OSAResponseDTO> postContrib(@RequestHeader(value = "Authorization") String token,
-    @PathVariable("art_id") Integer artId, @RequestBody PostContribDTO contrib)
+    		@PathVariable("art_id") Integer artId, @RequestBody PostContribDTO contrib)
 			throws OSA401Exception, OSA400Exception {
 
 		String username = jwtUtil.getUsernameFromToken(token.substring("Bearer ".length()));
@@ -68,8 +68,17 @@ public class ContribController {
 
     @DeleteMapping(value = "/contrib/{contrib_id}")
 	public ResponseEntity<OSAResponseDTO> deleteContrib(@RequestHeader(value = "Authorization") String token,
-														@PathVariable("contrib_id") Integer contribId) throws OSA401Exception, OSA404Exception {
-		// TODO : seulement le user qui a post'e contrib peut la delelte
+			@PathVariable("contrib_id") Integer contribId) throws OSA401Exception, OSA404Exception, OSA400Exception {
+
+		String username = jwtUtil.getUsernameFromToken(token.substring("Bearer ".length()));
+		Optional<UserEntity> optionalUser = userRepo.findByUsername(username);
+		if (!optionalUser.isPresent())
+			throw new OSA400Exception("No user found.");
+
+		Optional<ContribEntity> contrib = contribRepo.findById(contribId);
+		if (!contrib.get().getContributor().equals(optionalUser.get())
+				&& !jwtService.getRolesByToken(token.substring("Bearer ".length())).contains(RoleEnum.ROLE_ADMIN))
+			throw new OSA401Exception("Unauthorized.");										
 
 		contribService.delete(contribId);
 		return ResponseEntity.ok(new OSAResponseDTO("Contribution deleted"));
@@ -77,7 +86,8 @@ public class ContribController {
 
 	@PostMapping(value = "/contrib/accept/{contrib_id}")
 	public ResponseEntity<OSAResponseDTO> acceptContrib(@RequestHeader(value = "Authorization") String token,
-														@PathVariable("contrib_id") Integer contribId) throws OSA404Exception, OSA401Exception, OSA400Exception {
+			@PathVariable("contrib_id") Integer contribId) throws OSA404Exception, OSA401Exception, OSA400Exception {
+
 		if (!jwtService.getRolesByToken(token.substring("Bearer ".length())).contains(RoleEnum.ROLE_ADMIN))
 			throw new OSA401Exception("Unauthorized.");
 		
@@ -85,14 +95,14 @@ public class ContribController {
 		if(!contrib.isPresent())
 			throw new OSA404Exception("Contribution not found.");
 		
-		contribService.saveContrib(contrib.get());
+		contribService.acceptContrib(contrib.get());
     	return ResponseEntity.ok(new OSAResponseDTO("Contribution accepted"));
     }
 
 	@PostMapping(value = "/contrib/deny/{contrib_id}")
 	public ResponseEntity<OSAResponseDTO> denyContrib(@RequestHeader(value = "Authorization") String token,
-														@PathVariable("contrib_id") Integer contribId)
-			throws OSA401Exception, OSA404Exception, OSA400Exception {
+			@PathVariable("contrib_id") Integer contribId) throws OSA401Exception, OSA404Exception, OSA400Exception {
+
 		if (!jwtService.getRolesByToken(token.substring("Bearer ".length())).contains(RoleEnum.ROLE_ADMIN))
 			throw new OSA401Exception("Unauthorized.");
 
@@ -108,7 +118,7 @@ public class ContribController {
     public ResponseEntity<OSAResponseDTO> getContribs(@PathVariable("art_id") Integer artId) throws OSA404Exception {
 		Optional<ArtEntity> art = artRepo.findById(artId);
 		if(!art.isPresent())
-			throw new OSA404Exception("art not found.");
+			throw new OSA404Exception("Art not found.");
 
 		List<ContribEntity> contribs = new ArrayList<>(contribRepo.findByArtId(artId));
